added settings with email & phone contact option
