+++ app/src/main/AndroidManifest.xml
@@ -15,15 +15,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Oberesultanalysismobileapp1"
         tools:targetApi="31">
-        <activity
-            android:name=".FirebaseSignInActivity"
-            android:exported="true"
-            android:theme="@style/Theme.Oberesultanalysismobileapp1.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
+<!--        <activity-->
+<!--            android:name=".FirebaseSignInActivity"-->
+<!--            android:exported="true"-->
+<!--            android:theme="@style/Theme.Oberesultanalysismobileapp1.NoActionBar">-->
+<!--            <intent-filter>-->
+<!--                <action android:name="android.intent.action.MAIN" />-->
+<!--                <category android:name="android.intent.category.LAUNCHER" />-->
+<!--            </intent-filter>-->
+<!--        </activity>-->
         <activity
             android:name=".MainActivity"
             android:exported="true"
@@ -31,7 +31,7 @@
             android:theme="@style/Theme.Oberesultanalysismobileapp1.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
+                <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
     </application>
