+++ app/src/main/res/values/colors.xml
@@ -17,4 +17,11 @@
     <color name="color5">#3f51b5</color>
     <color name="color6">#1a237e</color>
     <color name="backgroundTint">#e3f2fd</color>
+
+
+    <!-- Additional colors for dark theme -->
+    <color name="gray_300">#E0E0E0</color>
+    <color name="gray_700">#616161</color>
+    <color name="gray_800">#424242</color>
+    <color name="gray_900">#212121</color>
 </resources>
