+++ app/src/main/res/values/themes.xml
@@ -11,14 +11,23 @@
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
         <item name="android:statusBarColor">?attr/colorPrimaryVariant</item>
-        <!-- Customize your theme here. -->
+        <!-- Background & Surface colors -->
+        <item name="android:colorBackground">@color/white</item>
+        <item name="colorSurface">@color/white</item>
+        <item name="colorOnSurface">@color/black</item>
+        <!-- Text colors -->
+        <item name="android:textColorPrimary">@color/black</item>
+        <item name="android:textColorSecondary">@color/gray_700</item>
     </style>
+
     <style name="Theme.Oberesultanalysismobileapp1.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
+
     <style name="Theme.Oberesultanalysismobileapp1.AppBarOverlay"
         parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+
     <style name="Theme.Oberesultanalysismobileapp1.PopupOverlay"
         parent="ThemeOverlay.AppCompat.Light" />
 </resources>
